Allow ''timeout'' parameter to be set on sh:exec, python:exec and python:distutils commands, which kills the commands after the given number of seconds. This makes it much easier to deal with runaway processes.
Small fix to killing processes-- ''kill'' isn't an attribute of
subprocess. It's an attribute of subprocess.Popen.

This is a good solution for #380-- set limits on the processes running
the steps, and you'll eventually get through the build. Other tickets
for ignoring a build fill in the rest. Closes #380.

Will update http://bitten.edgewall.org/wiki/Documentation/commands.html when
this goes in.