Add new last_activity field to build. I considered reusing stopped, but this seemed cleaner and more obvious, which seems like the right way to go.
On the completion of every step, last activity is updated to the current time.
All orphaning/invalidation based on slave_timeout happens based on that
time, not started. Previously, builds were orphaned if it had been more
than slave_timeout (config setting) seconds since they had started, now
it's since the last interaction (which is actually what the documentation
on the field already says.)

This requires a database upgrade.

Thanks to Hodgestar for comments.

Closes #222. Refs #411.